Implemented the enemy Klingon Bird of Prey

+++ SETTrek/SETTrek/Graphics.h
@@ -26,7 +26,7 @@ using namespace Microsoft::WRL;	// For using ComPtr
 //=============================
 // GRAPHIC CONSTANTS
 //=============================
-#define PI                      (float)3.14159265
+#define PI                      3.14159265f
 #define kDefaultScreenWidth		1024
 #define kDefaultScreenHeight	768
